requirements.txt
openai
langchain
unstructured
pdfminer.six
pdf2image
tiktoken
pinecone-client